Move loss group testing under tests/engine to better match package structure
